+++ engine/runtime/gltf/src/next/glb/mod.rs
@@ -11,6 +11,7 @@
 //! layer that builds Cesium's metadata extensions on top of this.
 
 mod primitive;
+mod texture;
 
 use std::borrow::Cow;
 use std::collections::{BTreeMap, HashSet};
@@ -19,6 +20,7 @@ use gltf::json;
 use gltf::json::validation::{Checked, USize64};
 
 pub use primitive::{normal, texcoord, DedupAttribute, Granularity};
+pub use texture::{ImageRef, MagFilter, MinFilter, SamplerDesc, TextureRef, Wrap};
 
 /// A material's PBR metallic-roughness description: the base factors, plus an
 /// optional base-color texture. A textured material's primitive must also carry
@@ -29,48 +31,6 @@ pub struct MaterialDesc {
     pub roughness_factor: f32,
     /// `baseColorTexture`; `None` for a colour-only material.
     pub base_color_texture: Option<TextureRef>,
-}
-
-/// Handle to a texture pushed via [`Builder::push_image_texture`], for
-/// referencing from a [`MaterialDesc`].
-#[derive(Clone, Copy)]
-pub struct TextureRef(json::Index<json::Texture>);
-
-impl TextureRef {
-    pub(super) fn index(self) -> json::Index<json::Texture> {
-        self.0
-    }
-}
-
-/// How a texture's coordinates are wrapped and filtered — the glTF-agnostic
-/// subset this writer needs. Maps to a glTF `sampler`.
-#[derive(Clone, Copy)]
-pub struct SamplerDesc {
-    pub wrap_s: Wrap,
-    pub wrap_t: Wrap,
-    pub mag: MagFilter,
-    pub min: MinFilter,
-}
-
-#[derive(Clone, Copy)]
-pub enum Wrap {
-    Repeat,
-    MirroredRepeat,
-    ClampToEdge,
-}
-
-#[derive(Clone, Copy)]
-pub enum MagFilter {
-    Nearest,
-    Linear,
-}
-
-#[derive(Clone, Copy)]
-pub enum MinFilter {
-    Nearest,
-    Linear,
-    NearestMipmap,
-    LinearMipmap,
 }
 
 /// Opaque handle to a primitive pushed via [`Builder::push_primitive`], for
@@ -151,50 +111,12 @@ impl Builder {
         })
     }
 
-    /// Embed an image as a bufferView-backed glTF texture (self-contained: no
-    /// external URI), with its sampler, and return a handle for a
-    /// [`MaterialDesc::base_color_texture`]. `mime_type` is the image's MIME
-    /// (e.g. `"image/png"`).
-    pub fn push_image_texture(
-        &mut self,
-        image_bytes: &[u8],
-        mime_type: &str,
-        sampler: SamplerDesc,
-    ) -> TextureRef {
-        let buffer_view = self.push_buffer_view_targeted(image_bytes, None);
-        let image = self.root.push(json::Image {
-            name: None,
-            buffer_view: Some(buffer_view),
-            mime_type: Some(json::image::MimeType(mime_type.to_string())),
-            uri: None,
-            extensions: Default::default(),
-            extras: Default::default(),
-        });
-        let sampler_index = self.root.push(json::texture::Sampler {
-            name: None,
-            mag_filter: Some(Checked::Valid(match sampler.mag {
-                MagFilter::Nearest => json::texture::MagFilter::Nearest,
-                MagFilter::Linear => json::texture::MagFilter::Linear,
-            })),
-            min_filter: Some(Checked::Valid(match sampler.min {
-                MinFilter::Nearest => json::texture::MinFilter::Nearest,
-                MinFilter::Linear => json::texture::MinFilter::Linear,
-                MinFilter::NearestMipmap => json::texture::MinFilter::NearestMipmapNearest,
-                MinFilter::LinearMipmap => json::texture::MinFilter::LinearMipmapLinear,
-            })),
-            wrap_s: Checked::Valid(wrap(sampler.wrap_s)),
-            wrap_t: Checked::Valid(wrap(sampler.wrap_t)),
-            extensions: None,
-            extras: Default::default(),
-        });
-        let texture = self.root.push(json::Texture {
-            name: None,
-            sampler: Some(sampler_index),
-            source: image,
-            extensions: Default::default(),
-            extras: Default::default(),
-        });
-        TextureRef(texture)
+    /// Mark `name` in both `extensionsUsed` and `extensionsRequired`.
+    pub fn require_extension(&mut self, name: &'static str) {
+        self.mark_extension_used(name);
+        if !self.root.extensions_required.iter().any(|n| n == name) {
+            self.root.extensions_required.push(name.to_string());
+        }
     }
 
     /// Append a `VEC3` f32 accessor (positions or normals); `with_bounds`
@@ -420,14 +342,6 @@ impl Builder {
             bin: Some(Cow::Owned(self.bin)),
         };
         glb.to_vec().expect("GLB binary output is always writable")
-    }
-}
-
-fn wrap(w: Wrap) -> json::texture::WrappingMode {
-    match w {
-        Wrap::Repeat => json::texture::WrappingMode::Repeat,
-        Wrap::MirroredRepeat => json::texture::WrappingMode::MirroredRepeat,
-        Wrap::ClampToEdge => json::texture::WrappingMode::ClampToEdge,
     }
 }
